No offchain workers.
The immediate trigger for removing this code is that it no longer compiled. Of course I could make it compile by copying the new version from the node template.

The bigger reason I'm removing it is because the template doesn't use an offchain worker, and I don't intend to install one, and I'm not convinced that offchain workers are even a good idea to begin with.