Update extrinsics root in test
The transactions encode differently now becuase there is a Vec of peeks in it. Even if there are no peeks (like all of the pre-existing tests) the encoding still chenges because a Vec's encoding contains it's length, so even an empty (length-of-zero) Vec still inserts a byte into the encoding.