Implement Peeks (#49)
* Update the `Transaction` type

* first pass at executive

* `ConstraintChekcerTrait`

* Sketch new executive tests

* runtime compiles

* Amoeba tests compile

* Money tests compile

* cargo fmt

* `TestTransaction` builder pattern

* Slightly improve test externality builder pattern

* refactor executive tests and include new tests for peek

* update extrinsics root now that extrinsic format has changed

* trivial updates to wallet

* amoeba stuff leftover from merging main

* Update custom encode stuff

* Fix type

* update aggregate constraint checker macro

* Improve money tests

* Fix ordering: inputs, peeks, outputs

* Update kitties tests

* fix name ordering in macro

This only makes  the variable names match the contents. It does not change behavior in any way.

* Update extrinsics root in test

The transactions encode differently now becuase there is a Vec of peeks in it. Even if there are no peeks (like all of the pre-existing tests) the encoding still chenges because a Vec's encoding contains it's length, so even an empty (length-of-zero) Vec still inserts a byte into the encoding.

* Remove inherited hack in clone no bound.

I'm not sure what "some reason" was, but `cargo fmt` seems to be working fine for me.

---------

Co-authored-by: Joshy Orndorff <git-user-email.h0ly5@simplelogin.com>