tried to write some light tests for `validate_block` but it isn't working.
This is tricky because the `validate_block` function is only built to wasm. I think it has to be that way because overwriting host functions... Maybe there is more to be explored here, but I found it not that rewarding of a direction. Feel free to explore more especially if we get desparate.

I'll revert this commit, but I'm checking it in just so I can publish this somewhere easily.